1.74.7 (5.6.4f1)
Build-Id: 7d25f66b-ea86-48a5-be51-3dbe3982454e
Build-Sha256: 7d06c499c224353815e5d2157bdc9a2ff3b955cdd06736b762f5c231808f57f2

+++ Assembly-CSharp/Avatar.cs
@@ -158,7 +158,7 @@ public class Avatar : MonoBehaviour, IBulletImpactVisualizer
 			string text = "_" + key;
 			if (!newModifiers.ContainsKey(text))
 			{
-				AvatarModifierPackageType item = (AvatarModifierPackageType)(int)Enum.Parse(typeof(AvatarModifierPackageType), text.TrimStart('_'));
+				AvatarModifierPackageType item = (AvatarModifierPackageType)Enum.Parse(typeof(AvatarModifierPackageType), text.TrimStart('_'));
 				list.Add(item);
 			}
 		}
@@ -177,7 +177,7 @@ public class Avatar : MonoBehaviour, IBulletImpactVisualizer
 		foreach (KeyValuePair<object, object> newModifier in newModifiers)
 		{
 			string text2 = (string)newModifier.Key;
-			AvatarModifierPackageType avatarModifierPackageType2 = (AvatarModifierPackageType)(int)Enum.Parse(typeof(AvatarModifierPackageType), text2.TrimStart('_'));
+			AvatarModifierPackageType avatarModifierPackageType2 = (AvatarModifierPackageType)Enum.Parse(typeof(AvatarModifierPackageType), text2.TrimStart('_'));
 			if (!modifiers.ContainsKey(avatarModifierPackageType2))
 			{
 				AvatarModifier avatarModifier = AvatarModifier.CreateFromType(avatarModifierPackageType2, this);
